Create linux_gpu_prerequisites
This bash shell script builds ramjet requirements, including TensorFlow 2.0 for use on embedded GPU development boards. Having said this, the GPU development boards all run a variation on Ubuntu 18.04. this is not optimized at all. Using sudo in this way is not always recommended. However, it works on the embedded boards we're evaluating.